feat:#1-모듈화 진행
