+++ src/Title.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+//Title 컴포넌트
+const Title = () => {
+    return <h1>2024 파리 올림픽</h1>;
+}
+
+export default Title
